requirements.txt
click==8.1.3
git+https://github.com/rwalk/gsdmm.git#egg=gsdmm
gensim==4.2.0
wordcloud==1.8.2.2 
matplotlib==3.5.2
numpy==1.23.1
regex==2022.7.9
nltk==3.7
requests==2.28.1
fastapi==0.79.0
uvicorn==0.17.6
spacy==3.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.3.0/en_core_web_md-3.3.0-py3-none-any.whl
watchfiles==0.16.1
fastapi-utils==0.2.1
email-validator==1.2.1
gunicorn==20.1.0
pymediawiki==0.7.1
networkx==2.8.6 
scikit-learn==1.1.2
pygraphviz==1.10
